+++ src/bedrock/world/inventory/network/item_stack_request_action_handler.cpp
@@ -1,0 +1,23 @@
+// Copyright (c) 2024, The Endstone Project. (https://endstone.dev) All Rights Reserved.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#include "bedrock/world/inventory/network/item_stack_request_action_handler.h"
+
+#include "bedrock/symbol.h"
+
+std::shared_ptr<SimpleSparseContainer> ItemStackRequestActionHandler::_getOrInitSparseContainer(
+    const FullContainerName &open_container_id)
+{
+    return BEDROCK_CALL(&ItemStackRequestActionHandler::_getOrInitSparseContainer, this, open_container_id);
+}