+++ src/endstone/core/block/block_face.h
@@ -14,12 +14,15 @@
 
 #pragma once
 
+#include <optional>
+
 #include "endstone/block/block_face.h"
 
 namespace endstone::core {
 
 class EndstoneBlockFace {
 public:
+    static std::optional<BlockFace> fromBedrockFacing(int facing);
     static int getOffsetX(BlockFace face);
     static int getOffsetY(BlockFace face);
     static int getOffsetZ(BlockFace face);