+++ src/endstone/core/entity/components/flag_components.h
@@ -14,9 +14,26 @@
 
 #pragma once
 
+#include <optional>
+
+#include "bedrock/world/level/block_pos.h"
+#include "endstone/event/block/cauldron_level_change_event.h"
 #include "endstone/event/player/player_open_sign_event.h"
+#include "endstone/inventory/item_stack.h"
+
+class Actor;
 
 namespace endstone::core {
+struct InternalBucketFillActorComponent {
+    const ::Actor *target = nullptr;
+    std::optional<ItemStack> item_stack;
+    bool write_item_stack = false;
+};
+struct InternalCauldronInteractionComponent {
+    Nullable<Actor> actor;
+    BlockPos pos;
+    CauldronLevelChangeEvent::ChangeReason reason = CauldronLevelChangeEvent::ChangeReason::Unknown;
+};
 struct InternalDisconnectFlagComponent {};
 struct InternalOpenSignCauseComponent {
     PlayerOpenSignEvent::Cause cause = PlayerOpenSignEvent::Cause::Unknown;